requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
requests==2.32.5
apache-airflow>=3.0.0
duckdb==1.1.0        
dbt-core>=1.7.0
dbt-duckdb==1.8.0